LayoutTests/platform/blackberry/editing/selection/click-in-margins-inside-editable-div-expected.txt

Ensure cursor placement matches IE6/IE7/FF3 when clicking above/below lines in padding.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true]
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 1 focusNode: [object Text](XX) focusOffset: 1 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 0 focusNode: [object Text] focusOffset: 0
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 0 focusNode: [object Text] focusOffset: 0
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true]
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 2 focusNode: [object Text] focusOffset: 2
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 1 focusNode: [object Text](XX) focusOffset: 1 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 2 focusNode: [object Text] focusOffset: 2
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true]
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 0 focusNode: [object Text] focusOffset: 0
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 1 focusNode: [object Text](XX) focusOffset: 1 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 0 focusNode: [object Text] focusOffset: 0
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 0 focusNode: [object Text] focusOffset: 0
FAIL Selection is [anchorNode: [object Text](YY) anchorOffset: 0 focusNode: [object Text](YY) focusOffset: 0 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 2 focusNode: [object Text] focusOffset: 2
FAIL Selection is [anchorNode: [object Text](YY) anchorOffset: 1 focusNode: [object Text](YY) focusOffset: 1 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 2 focusNode: [object Text] focusOffset: 2
PASS Selection is [anchorNode: [object Text](YY) anchorOffset: 2 focusNode: [object Text](YY) focusOffset: 2 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 1 focusNode: [object Text](XX) focusOffset: 1 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 1 focusNode: [object Text](XX) focusOffset: 1 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true]
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 0 focusNode: [object Text](XX) focusOffset: 0 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 0 focusNode: [object Text] focusOffset: 0
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 1 focusNode: [object Text](XX) focusOffset: 1 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 1 focusNode: [object Text] focusOffset: 1
FAIL Selection is [anchorNode: [object Text](XX) anchorOffset: 2 focusNode: [object Text](XX) focusOffset: 2 isCollapsed: true] should be at anchorNode: [object Text] anchorOffset: 2 focusNode: [object Text] focusOffset: 2
PASS Selection is [anchorNode: [object Text](YY) anchorOffset: 0 focusNode: [object Text](YY) focusOffset: 0 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](YY) anchorOffset: 1 focusNode: [object Text](YY) focusOffset: 1 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](YY) anchorOffset: 2 focusNode: [object Text](YY) focusOffset: 2 isCollapsed: true]
PASS successfullyParsed is true

TEST COMPLETE
